Move in-memory strategy to support folder
