+++ spec/upperkut/processor_spec.rb
@@ -1,44 +1,14 @@
 require 'spec_helper'
-require 'upperkut/util'
+require 'support/in_memory_strategy'
 require 'upperkut/processor'
 
 module Upperkut
   RSpec.describe Processor do
     subject(:processor) { described_class.new(worker, logger) }
 
+    let(:strategy) { worker.strategy }
     let(:worker) { DummyWorker }
     let(:logger) { Logger.new(nil) }
-
-    class InMemoryStrategy
-      include Util
-
-      def initialize
-        @items = []
-      end
-
-      def push_items(items)
-        @items.concat(normalize_items(items))
-      end
-
-      def fetch_items
-        @items.slice!(0..10)
-      end
-
-      def clear
-        @items.clear
-      end
-
-      def process?
-        true
-      end
-
-      def metrics
-        {
-          'size' => @items.size,
-          'latency' => Time.now.to_i - @items.first.enqueued_at
-        }
-      end
-    end
 
     class DummyWorker
       include Worker
